ai-engine/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
openai>=1.30.0
tiktoken>=0.7.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
duckduckgo-search>=6.0.0
mysql-connector-python>=9.0.0
simpleeval>=1.0.0
